refactor: String 타입이 아니면 Column length 가 sql 에 반영되지 않도록 수정

+++ src/main/java/persistence/sql/ddl/TypeMapper.java
@@ -6,7 +6,7 @@ import java.util.Map;
 public final class TypeMapper {
     private static final Map<Class<?>, String> javaToSql = Map.of(
             Long.class, "BIGINT",
-            Integer.class, "BIGINT",
+            Integer.class, "INTEGER",
             String.class, "VARCHAR"
     );
